Move to BigDecimal amounts

+++ app/models/Operation.java
@@ -6,6 +6,7 @@ import play.db.jpa.Model;
 
 import javax.persistence.Entity;
 import javax.persistence.ManyToOne;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.UUID;
 
@@ -15,19 +16,22 @@ public class Operation extends Model {
     public String uniq;
     @Required
     @Min(0.01)
-    public Float amount;
+    public BigDecimal amount;
+
     @Required
     public Date date;
+
     @ManyToOne
     @Required
     public OperationType operationType;
+
     @Required
     @ManyToOne
     public User user;
 
     public Operation() { uniq = UUID.randomUUID().toString(); }
 
-    public Operation(Float amount, Date date, OperationType operationType, User user) {
+    public Operation(BigDecimal amount, Date date, OperationType operationType, User user) {
         uniq = UUID.randomUUID().toString();
         this.amount = amount;
         this.date = date;
